+++ minimal.rb
@@ -107,6 +107,7 @@ after_bundle do
 
   apply "src/shared/env_rubocop.rb"
   apply "src/shared/migrations_uuid.rb"
+  apply "src/shared/config_app_settings.rb"
 
   apply "src/shared/gems_rspec.rb"
   apply "src/shared/gems_i18n_tasks.rb"