+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [better_html](https://github.com/Shopify/better-html) gem with spec
 - [erb_lint](https://github.com/Shopify/erb_lint) gem
+- custom settings file - `config/settings.yml`
 
 ## [1.0.1]